Build: grunt-jquery-content 3.1 to 3.3, cheerio 0.22 to 1.0
It seems that the cheerio upgrade does not require changes in this
repository. When we upgraded it in grunt-jquery-content, we needed
https://github.com/jquery/grunt-jquery-content/commit/72a287e903.

Here, the HTML files loaded in cheerio are from the demos directory,
which are full HTML documents including `<html>`. I compared the
dist/ directory before and after this to make sure, and while there
are minor whitespace and character encoding differences, there are no
structural differences that I could find.